# apps/api/library-mini-admin-api/src/main/generated/src/main/java/com/example/library/generated/api/BooksApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.25.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.library.generated.api;

import com.example.library.generated.dto.ErrorResponseDTO;
import com.example.library.generated.dto.GetBooksResponseDTO;
import com.example.library.generated.dto.PostBooksRequestDTO;
import com.example.library.generated.dto.PostBooksResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-09-07T10:06:39.879502+08:00[Asia/Taipei]", comments = "Generator version: 7.25.0")
@Validated
@Tag(name = "Books", description = "Catalogue and inventory operations")
public interface BooksApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_BOOKS = "/api/books";
    /**
     * GET /api/books : [library-books-001] 取得館藏列表
     *
     * @return 館藏列表查詢成功 (status code 200)
     *         or 館藏列表查詢失敗 (status code 500)
     */
    @Operation(
        operationId = "getBooks",
        summary = "[library-books-001] 取得館藏列表",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "200", description = "館藏列表查詢成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = GetBooksResponseDTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "館藏列表查詢失敗", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = BooksApi.PATH_GET_BOOKS,
        produces = { "application/json" }
    )
    default ResponseEntity<GetBooksResponseDTO> getBooks(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"data\" : [ { \"bookId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"title\" : \"title\", \"isbn\" : \"isbn\", \"author\" : \"author\", \"category\" : \"category\", \"status\" : \"AVAILABLE\", \"availableCount\" : 0, \"totalCount\" : 1, \"isActive\" : true }, { \"bookId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"title\" : \"title\", \"isbn\" : \"isbn\", \"author\" : \"author\", \"category\" : \"category\", \"status\" : \"AVAILABLE\", \"availableCount\" : 0, \"totalCount\" : 1, \"isActive\" : true } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"details\" : [ { \"field\" : \"field\", \"reason\" : \"reason\" }, { \"field\" : \"field\", \"reason\" : \"reason\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_POST_BOOKS = "/api/books";
    /**
     * POST /api/books : [library-books-002] 建立書籍與初始館藏
     *
     * @param postBooksRequestDTO  (required)
     * @return 書籍建立成功 (status code 201)
     *         or 輸入資料或唯一性條件不符合 (status code 400)
     *         or 書籍建立失敗 (status code 500)
     */
    @Operation(
        operationId = "postBooks",
        summary = "[library-books-002] 建立書籍與初始館藏",
        tags = { "Books" },
        responses = {
            @ApiResponse(responseCode = "201", description = "書籍建立成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PostBooksResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "輸入資料或唯一性條件不符合", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponseDTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "書籍建立失敗", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = BooksApi.PATH_POST_BOOKS,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<PostBooksResponseDTO> postBooks(
        @Parameter(name = "PostBooksRequestDTO", description = "", required = true) @Valid @RequestBody PostBooksRequestDTO postBooksRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"data\" : { \"bookId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"title\" : \"title\", \"isbn\" : \"isbn\", \"author\" : \"author\", \"category\" : \"category\", \"status\" : \"AVAILABLE\", \"availableCount\" : 0, \"totalCount\" : 1, \"isActive\" : true } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"details\" : [ { \"field\" : \"field\", \"reason\" : \"reason\" }, { \"field\" : \"field\", \"reason\" : \"reason\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"details\" : [ { \"field\" : \"field\", \"reason\" : \"reason\" }, { \"field\" : \"field\", \"reason\" : \"reason\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}