# apps/api/library-mini-admin-api/src/main/generated/src/main/java/com/example/library/generated/api/LoansApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.25.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.example.library.generated.api;

import com.example.library.generated.dto.ErrorResponseDTO;
import com.example.library.generated.dto.PostLoansBorrowRequestDTO;
import com.example.library.generated.dto.PostLoansBorrowResponseDTO;
import com.example.library.generated.dto.PostLoansReturnRequestDTO;
import com.example.library.generated.dto.PostLoansReturnResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-09-07T10:06:39.879502+08:00[Asia/Taipei]", comments = "Generator version: 7.25.0")
@Validated
@Tag(name = "Loans", description = "Borrow and return operations")
public interface LoansApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_POST_LOANS_BORROW = "/api/loans/borrow";
    /**
     * POST /api/loans/borrow : [library-loans-001] 借出書籍
     *
     * @param postLoansBorrowRequestDTO  (required)
     * @return 書籍借出成功 (status code 200)
     *         or 書籍不可借或借閱條件不符合 (status code 400)
     *         or 借出處理失敗 (status code 500)
     */
    @Operation(
        operationId = "postLoansBorrow",
        summary = "[library-loans-001] 借出書籍",
        tags = { "Loans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "書籍借出成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PostLoansBorrowResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "書籍不可借或借閱條件不符合", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponseDTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "借出處理失敗", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoansApi.PATH_POST_LOANS_BORROW,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<PostLoansBorrowResponseDTO> postLoansBorrow(
        @Parameter(name = "PostLoansBorrowRequestDTO", description = "", required = true) @Valid @RequestBody PostLoansBorrowRequestDTO postLoansBorrowRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"data\" : { \"loanId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"bookId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"readerId\" : \"readerId\", \"borrowedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"dueDate\" : \"2000-01-23\", \"returnedAt\" : \"2000-01-23T04:56:07.000+00:00\" } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"details\" : [ { \"field\" : \"field\", \"reason\" : \"reason\" }, { \"field\" : \"field\", \"reason\" : \"reason\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"details\" : [ { \"field\" : \"field\", \"reason\" : \"reason\" }, { \"field\" : \"field\", \"reason\" : \"reason\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_POST_LOANS_RETURN = "/api/loans/return";
    /**
     * POST /api/loans/return : [library-loans-002] 歸還書籍
     *
     * @param postLoansReturnRequestDTO  (required)
     * @return 書籍歸還成功 (status code 200)
     *         or 沒有可歸還借閱或歸還條件不符合 (status code 400)
     *         or 歸還處理失敗 (status code 500)
     */
    @Operation(
        operationId = "postLoansReturn",
        summary = "[library-loans-002] 歸還書籍",
        tags = { "Loans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "書籍歸還成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PostLoansReturnResponseDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "沒有可歸還借閱或歸還條件不符合", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponseDTO.class))
            }),
            @ApiResponse(responseCode = "500", description = "歸還處理失敗", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = LoansApi.PATH_POST_LOANS_RETURN,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<PostLoansReturnResponseDTO> postLoansReturn(
        @Parameter(name = "PostLoansReturnRequestDTO", description = "", required = true) @Valid @RequestBody PostLoansReturnRequestDTO postLoansReturnRequestDTO
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"00000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"data\" : { \"loanId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"bookId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"readerId\" : \"readerId\", \"borrowedAt\" : \"2000-01-23T04:56:07.000+00:00\", \"dueDate\" : \"2000-01-23\", \"returnedAt\" : \"2000-01-23T04:56:07.000+00:00\" } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"details\" : [ { \"field\" : \"field\", \"reason\" : \"reason\" }, { \"field\" : \"field\", \"reason\" : \"reason\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : \"A0000\", \"message\" : \"message\", \"traceId\" : \"traceId\", \"details\" : [ { \"field\" : \"field\", \"reason\" : \"reason\" }, { \"field\" : \"field\", \"reason\" : \"reason\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}